+++ application/src/main/java/org/thingsboard/server/service/security/auth/TokenOutdatingService.java
@@ -15,46 +15,12 @@
  */
 package org.thingsboard.server.service.security.auth;
 
-import io.jsonwebtoken.Claims;
-import lombok.RequiredArgsConstructor;
-import org.springframework.context.event.EventListener;
-import org.springframework.stereotype.Service;
-import org.springframework.util.StringUtils;
-import org.thingsboard.server.cache.TbTransactionalCache;
 import org.thingsboard.server.common.data.id.UserId;
 import org.thingsboard.server.common.data.security.event.UserAuthDataChangedEvent;
 import org.thingsboard.server.common.data.security.model.JwtToken;
-import org.thingsboard.server.service.security.model.token.JwtTokenFactory;
 
-import java.util.Optional;
+public interface TokenOutdatingService {
+    void onUserAuthDataChanged(UserAuthDataChangedEvent event);
 
-import static java.util.concurrent.TimeUnit.MILLISECONDS;
-
-@Service
-@RequiredArgsConstructor
-public class TokenOutdatingService {
-    private final TbTransactionalCache<String, Long> cache;
-    private final JwtTokenFactory tokenFactory;
-
-    @EventListener(classes = UserAuthDataChangedEvent.class)
-    public void onUserAuthDataChanged(UserAuthDataChangedEvent event) {
-        if (StringUtils.hasText(event.getId())) {
-            cache.put(event.getId(), event.getTs());
-        }
-    }
-
-    public boolean isOutdated(JwtToken token, UserId userId) {
-        Claims claims = tokenFactory.parseTokenClaims(token).getBody();
-        long issueTime = claims.getIssuedAt().getTime();
-        String sessionId = claims.get("sessionId", String.class) == null ? "" : claims.get("sessionId", String.class);
-        return isTokenOutdated(issueTime, userId.toString()) || isTokenOutdated(issueTime, sessionId);
-    }
-
-    private Boolean isTokenOutdated(long issueTime, String sessionId) {
-        return Optional.ofNullable(cache.get(sessionId)).map(outdatageTime -> isTokenOutdated(issueTime, outdatageTime.get())).orElse(false);
-    }
-
-    private boolean isTokenOutdated(long issueTime, Long outdatageTime) {
-        return MILLISECONDS.toSeconds(issueTime) < MILLISECONDS.toSeconds(outdatageTime);
-    }
+    boolean isOutdated(JwtToken token, UserId userId);
 }
